+++ packages/wm/src/common/mod.rs
@@ -5,12 +5,12 @@ mod display_state;
 pub mod events;
 mod length_value;
 mod memo;
+mod opacity_value;
 pub mod platform;
 mod point;
 mod rect;
 mod rect_delta;
 mod tiling_direction;
-mod transparency_value;
 mod try_warn;
 mod vec_deque_ext;
 
@@ -19,9 +19,9 @@ pub use direction::*;
 pub use display_state::*;
 pub use length_value::*;
 pub use memo::*;
+pub use opacity_value::*;
 pub use point::*;
 pub use rect::*;
 pub use rect_delta::*;
 pub use tiling_direction::*;
-pub use transparency_value::*;
 pub use vec_deque_ext::*;
